1.refactoring: taking some functions and bring them to app.jsx to send results as props for navbar and index.jsx 2. adding toFixed() to amounts on index.jsx
