+++ src/App.css
@@ -1,6 +1,4 @@
 .full-container {
     display: grid;
     align-self: center;
-    
 }
-
