+++ src/App.jsx
@@ -7,17 +7,35 @@ import Edit from "./Pages/Edit";
 import New from "./Pages/New";
 import NotFound from "./Pages/NotFound";
 import NavBar from './Components/NavBar';
+import { useState, useEffect } from "react";
 
 
 function App() {
 
+    const [transactions, setTransactions] = useState([]);
+
+    const API = import.meta.env.VITE_API_URL;
+
+    useEffect(() => {
+        fetch(`${API}/transactions`)
+        .then(res => res.json())
+        .then(resJSON => setTransactions(resJSON))
+        .catch((error) => {
+            console.error(error);
+        });
+    }, []);
+
+  const incomeSum = transactions.filter(trans => trans.category === "Income").reduce((sum, {amount}) => sum + +amount, 0);
+  const expenseSum = transactions.filter(trans => trans.category === "Expense").reduce((sum, {amount}) => sum - +amount, 0);
+  const balance = incomeSum + expenseSum;
+
   return (
     <>
       <Router>
-        <NavBar />
+        <NavBar balance={balance}/>
         <Routes>
           <Route path='/' element={ <Home/> } />
-          <Route path='/transactions' element={ <Index /> } />
+          <Route path='/transactions' element={ <Index incomeSum={incomeSum} expenseSum={expenseSum} /> } />
           <Route path='/transactions/:id' element={ <Show /> } />
           <Route path='/transactions/:id/edit' element={ <Edit /> } />
           <Route path='/transactions/new' element={ <New /> } />
