+++ src/Components/Transaction.jsx
@@ -8,7 +8,7 @@ export default function Transaction ({trans}) {
             <div className="transaction-container">
                 <Link to={`/transactions/${id}`}><p className="transaction-title">{transactionName}</p></Link>
                 <p className="transaction-info">{date}</p>
-                <p className="transaction-info">{category === "Expense" ? `-${amount}` : `+${amount}`}$</p>
+                <p className="transaction-info">{category === "Expense" ? `-${amount.toFixed(2)}` : `+${amount.toFixed(2)}`}$</p>
             </div>
         </>
     );
