+++ src/Pages/Home.jsx
@@ -6,7 +6,7 @@ export default function Home () {
             <div className="home-container">
                 <h1 className="app-title">Budgetmin</h1>
                 <p>"Master Your Money, Simplify Your Life"</p>
-                <img src="/benjamin.png" alt="golden coin" width="300" height="340"/>
+                <img src="/benjamin.png" alt="benjamin franklin portrait" width="300" height="340"/>
                 <Link to={"/transactions"}>
                     <button>Start</button>
                 </Link>
